% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdp_mcmc_object.R
\name{diagnose}
\alias{diagnose}
\title{Diagnose GPDPQuantReg}
\usage{
diagnose(x, ...)
}
\arguments{
\item{GPDP_MCMC}{A GPDP_MCMC object}
}
\value{
plots traces, autocorrelations, crosscorrelations and ergodicity
}
\description{
Run diagnostics for the GPDPQuantReg's Monte Carlo Markov Chains
}
\examples{
m <- 35
x <- sort(sample(seq(-15, 15, 0.005), m))
f_x <- function(x) return((1/40) * x^2 - (1/20) * x - 2)
data <- data.frame(x = x, y = f_x(x) + rnorm(m, 0, 1))
GPDP_MCMC <- fit_GPDPQuantReg(y ~ x, data, p = 0.250)
diagnose(GPDP_MCMC)
}
\author{
Omar Pardo (omarpardog@gmail.com)
}
