% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict.GPDP_MCMC}
\alias{predict.GPDP_MCMC}
\title{Predicts GPDPQuantReg}
\usage{
\method{predict}{GPDP_MCMC}(GPDP_MCMC, predictive_data, credibility = 0.95)
}
\arguments{
\item{GPDP_MCMC}{A GPDP_MCMC object}

\item{predictive_data}{data frame with the new X values, for prediction}

\item{credibility}{real between (0,1), size of the prediction's credible interval}
}
\value{
prediction data frame with the mean, median, lower and upper values for each x
}
\description{
Predicts after fitting a GPDPQuantReg
}
\examples{
m <- 35
x <- sort(sample(seq(-15, 15, 0.005), m))
f_x <- function(x) return((1/40) * x^2 - (1/20) * x - 2)
data <- data.frame(x = x, y = f_x(x) + rnorm(m, 0, 1))
GPDP_MCMC <- fit_GPDPQuantReg(y ~ x, data, p = 0.250)
predictive_data <- data_frame(x = seq(-15, 15, 0.1))
credibility <- 0.90
predict(GPDP_MCMC, predictive_data, credibility)
}
\author{
Omar Pardo (omarpardog@gmail.com)
}
